+++ src/result/LineSet.php
@@ -9,14 +9,15 @@
  * with this source code in the file LICENSE.
  */
 
-namespace cloak\value;
+namespace cloak\result;
 
+use cloak\value\Coverage;
+use cloak\value\LineRange;
 use PhpCollection\Sequence;
-use cloak\result\Line;
 
 /***
  * Class LineSet
- * @package cloak\value
+ * @package cloak\result
  */
 class LineSet
 {
@@ -101,6 +102,40 @@ class LineSet
     public function isCoverageGreaterEqual(Coverage $coverage)
     {
         return $this->getCodeCoverage()->greaterEqual($coverage);
+    }
+
+    /**
+     * @param LineRange $lineRange
+     * @return LineSet
+     */
+    public function selectRange(LineRange $lineRange)
+    {
+        $lines = $this->selectLines(function(Line $line) use ($lineRange) {
+            $lineNumber = $line->getLineNumber();
+            return $lineRange->contains($lineNumber);
+        });
+
+        return new self($lines);
+    }
+
+    /**
+     * @return null|Line
+     */
+    public function first()
+    {
+        $line = $this->lines->first();
+        $line = $line->isDefined() ? $line->get() : null;
+        return $line;
+    }
+
+    /**
+     * @return null|Line
+     */
+    public function last()
+    {
+        $line = $this->lines->last();
+        $line = $line->isDefined() ? $line->get() : null;
+        return $line;
     }
 
     /**